requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
cryptography==41.0.7
Werkzeug==2.3.7
click==8.1.7
nacos-sdk-python==0.1.12
PyYAML==6.0.1 